+++ components/script/dom/imagedata.rs
@@ -149,8 +149,8 @@ impl ImageData {
         }
     }
 
-    pub fn get_size(&self) -> Size2D<i32> {
-        Size2D::new(self.Width() as i32, self.Height() as i32)
+    pub fn get_size(&self) -> Size2D<u32> {
+        Size2D::new(self.Width(), self.Height())
     }
 }
